secure_channel/requirements.txt
websockets==11.0.3
pyjwt==2.8.0
cryptography==41.0.4  # For SSL/TLS support
python-dotenv==1.0.0
pyOpenSSL==23.2.0  # For SSL certificate generation